intro to jdbcTemplate's DataAccessException
JdbcTemplate wraps the check exception SQLException as an exception in
the runtime exception DataAccessException hierarchy, so the add ()
method does not need any code to wrap the exception. In addition,
regardless of the type of DB, errors caused by duplicate keys are thrown
by mapping to DuplicateKeyException, a subclass of DataAccessException.